feat(api): removing prisma dependency for render Service to be used in DocGen
