+++ api/src/core/render/render.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { RenderService } from './render.service';
+
+@Module({
+  providers: [RenderService]
+})
+export class RenderModule {}
