+++ api/src/core/template/template.controller.ts
@@ -1,149 +1,29 @@
 import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
-import { Template, TemplateType, Prisma } from '@prisma/client';
-import { PrismaService } from '../../prisma.service';
-import { TransformerService } from '../transformer/transformer.service';
+import { Template, Prisma } from '@prisma/client';
 import { RenderDto, RenderDtoTest, RenderResponse } from '../dto/render';
-import { JsTLService } from 'src/engines/jstl/jstl.service';
 import { TemplateService } from './template.service';
-import { JinjaService } from 'src/engines/jinja/jinja.service';
-import { EjsService } from 'src/engines/ejs/ejs.service';
+import { RenderService } from '../render/render.service';
 
 @Controller('/')
 export class TemplateController {
   constructor(
-    private readonly jinjaService: JinjaService,
-    private readonly ejsService: EjsService,
-    private readonly prisma: PrismaService,
-    private readonly transformerService: TransformerService,
     private readonly templateService: TemplateService,
-    private readonly jstlService: JsTLService,
+    private readonly renderService: RenderService,
   ) {}
 
   @Post('/process')
   async render(@Body() renderDto: RenderDto): Promise<RenderResponse> {
-    const template = await this.prisma.template.findUnique({
-      where: { id: renderDto.id },
-      include: {
-        transformers: {
-          include: {
-            transformer: true,
-          },
-        },
-      },
-    });
-    let processed;
-
-    let transformedData = renderDto.data;
-    for (const transformer of template.transformers) {
-      transformedData = await this.transformerService.process(
-        transformer.transformer,
-        transformedData,
-        transformer.path,
-      );
-    }
-    switch (template.type) {
-      case TemplateType.JINJA:
-        processed = this.jinjaService.render(template.body, transformedData);
-        break;
-
-      case TemplateType.JS_TEMPLATE_LITERALS:
-        processed = this.jstlService.render(template.body, transformedData);
-        break;
-
-      case TemplateType.EJS:
-        processed = this.ejsService.render(template.body, transformedData);
-        break;
-      default:
-        throw 'Templates without template types not allowed';
-    }
-    return {
-      processed,
-      templateType: template.type,
-      data: renderDto.data,
-      template: 'test',
-      meta: template.meta,
-    };
+    return this.renderService.renderTemplate(renderDto);
   }
 
   @Post('/process/test')
   async renderTest(@Body() data: RenderDtoTest): Promise<RenderResponse> {
-    let processed;
-    let transformedData;
-    try {
-      transformedData = JSON.parse(data.sampleData);
-    } catch (e) {
-      transformedData = data.sampleData;
-    }
-    // for (const transformer of template.transformers) {
-    //   transformedData = await this.transformerService.process(
-    //     transformer.transformer,
-    //     transformedData,
-    //     transformer.path,
-    //   );
-    // }
-    switch (data.type) {
-      case TemplateType.JINJA:
-        processed = this.jinjaService.render(data.body, transformedData);
-
-      case TemplateType.JS_TEMPLATE_LITERALS:
-        processed = this.jstlService.render(data.body, transformedData);
-        break;
-
-      case TemplateType.EJS:
-        processed = this.ejsService.render(data.body, transformedData);
-        break;
-      default:
-        throw 'Templates without template types not allowed';
-    }
-    return {
-      processed,
-      templateType: TemplateType.JS_TEMPLATE_LITERALS,
-      data: data.sampleData,
-      template: 'test',
-    };
+    return this.renderService.renderTemplateTest(data);
   }
 
   @Post('/process/testMany')
   async renderTestMany(@Body() data: RenderDtoTest): Promise<RenderResponse> {
-    const processed = [];
-    let transformedData;
-    try {
-      transformedData = JSON.parse(data.sampleData);
-    } catch (e) {
-      transformedData = data.sampleData;
-    }
-    for (let i = 0; i < transformedData.length; i++) {
-      switch (data.type) {
-        case TemplateType.JINJA:
-          processed.push({
-            __index: transformedData[i].__index,
-            body: this.jinjaService.render(data.body, transformedData[i]),
-          });
-
-        case TemplateType.JS_TEMPLATE_LITERALS:
-          processed.push({
-            __index: transformedData[i].__index,
-            body: this.jstlService.render(data.body, transformedData[i]),
-          });
-
-          break;
-
-        case TemplateType.EJS:
-          processed.push({
-            __index: transformedData[i].__index,
-            body: this.ejsService.render(data.body, transformedData[i]),
-          });
-          break;
-        default:
-          throw 'Templates without template types not allowed';
-      }
-    }
-    return {
-      processed,
-      templateType: TemplateType.JS_TEMPLATE_LITERALS,
-      data: data.sampleData,
-      template: 'test',
-    };
+    return this.renderService.renderTemplateManyTest(data);
   }
 
   @Post('/')
@@ -160,6 +40,6 @@ export class TemplateController {
 
   @Get('/:id')
   async getTemplate(@Param('id') id: string): Promise<Template> {
-    return this.templateService.template({ id: Number(id) });
+    return this.templateService.getTemplate({ id: Number(id) });
   }
 }
